# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API портала автоматизированного подсчёта строительных объёмов. Stage 1: проекты, импорт распознанных пакетов и просмотрщик документации. Расчёта объёмов и вызовов моделей на этом этапе нет.",
    "title": "Quantor API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/callback": {
      "get": {
        "description": "Меняет код на токены, заводит сеанс и возвращает браузер в портал.",
        "operationId": "complete_login",
        "parameters": [
          {
            "description": "Код авторизации провайдера",
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "description": "Код авторизации провайдера",
              "title": "Code",
              "type": "string"
            }
          },
          {
            "description": "Значение состояния",
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "description": "Значение состояния",
              "title": "State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "307": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Завершить вход",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "get": {
        "description": "Переадресует на страницу входа провайдера.\n\nВ dev-режиме провайдера нет: браузер сразу возвращается в портал, где его уже ждёт\nфиксированная личность.",
        "operationId": "begin_login",
        "parameters": [
          {
            "description": "Путь внутри портала",
            "in": "query",
            "name": "next",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Путь внутри портала",
              "title": "Next"
            }
          }
        ],
        "responses": {
          "307": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Начать вход",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}